requirements.txt
baidu-aip==2.2.18.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Cython==0.29.14
Flask==2.0.1
Flask-Cors==3.0.10
gensim==3.8.3
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
pycparser==2.20
pymongo==3.11.4
pypinyin==0.41.0
redis==3.5.3
requests==2.25.1
scipy==1.6.3
six==1.16.0
smart-open==5.1.0
urllib3==1.26.5
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
